+++ conn_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/attribute"
-	semconv "go.opentelemetry.io/otel/semconv/v1.7.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -137,10 +137,11 @@ var (
 
 func TestOtConn_Ping(t *testing.T) {
 	testCases := []struct {
-		name         string
-		error        bool
-		pingOption   bool
-		noParentSpan bool
+		name             string
+		error            bool
+		pingOption       bool
+		noParentSpan     bool
+		attributesGetter AttributesGetter
 	}{
 		{
 			name:       "ping enabled",
@@ -159,6 +160,10 @@ func TestOtConn_Ping(t *testing.T) {
 		{
 			name: "ping disabled",
 		},
+		{
+			name:             "with attribute getter",
+			attributesGetter: getDummyAttributesGetter(),
+		},
 	}
 
 	for _, tc := range testCases {
@@ -168,6 +173,7 @@ func TestOtConn_Ping(t *testing.T) {
 
 			// New conn
 			cfg.SpanOptions.Ping = tc.pingOption
+			cfg.AttributesGetter = tc.attributesGetter
 			mc := newMockConn(tc.error)
 			otelConn := newConn(mc, cfg)
 
@@ -189,9 +195,10 @@ func TestOtConn_Ping(t *testing.T) {
 						parentSpan:         dummySpan,
 						error:              tc.error,
 						expectedAttributes: cfg.Attributes,
-						expectedMethod:     MethodConnPing,
+						method:             MethodConnPing,
 						noParentSpan:       tc.noParentSpan,
 						ctx:                mc.pingCtx,
+						attributesGetter:   tc.attributesGetter,
 					})
 
 					assert.Equal(t, 1, mc.pingCount)
@@ -206,13 +213,17 @@ func TestOtConn_Ping(t *testing.T) {
 }
 
 func TestOtConn_ExecContext(t *testing.T) {
-	expectedAttrs := []attribute.KeyValue{semconv.DBStatementKey.String("query")}
+	query := "query"
+	args := []driver.NamedValue{{Value: "foo"}}
+	expectedAttrs := []attribute.KeyValue{semconv.DBStatementKey.String(query)}
+
 	testCases := []struct {
-		name         string
-		error        bool
-		noParentSpan bool
-		disableQuery bool
-		attrs        []attribute.KeyValue
+		name             string
+		error            bool
+		noParentSpan     bool
+		disableQuery     bool
+		attrs            []attribute.KeyValue
+		attributesGetter AttributesGetter
 	}{
 		{
 			name:  "no error",
@@ -232,6 +243,11 @@ func TestOtConn_ExecContext(t *testing.T) {
 			noParentSpan: true,
 			attrs:        expectedAttrs,
 		},
+		{
+			name:             "with attribute getter",
+			attributesGetter: getDummyAttributesGetter(),
+			attrs:            expectedAttrs,
+		},
 	}
 
 	for _, tc := range testCases {
@@ -241,10 +257,11 @@ func TestOtConn_ExecContext(t *testing.T) {
 
 			// New conn
 			cfg.SpanOptions.DisableQuery = tc.disableQuery
+			cfg.AttributesGetter = tc.attributesGetter
 			mc := newMockConn(tc.error)
 			otelConn := newConn(mc, cfg)
 
-			_, err := otelConn.ExecContext(ctx, "query", nil)
+			_, err := otelConn.ExecContext(ctx, query, args)
 			if tc.error {
 				require.Error(t, err)
 			} else {
@@ -260,9 +277,12 @@ func TestOtConn_ExecContext(t *testing.T) {
 				parentSpan:         dummySpan,
 				error:              tc.error,
 				expectedAttributes: append(cfg.Attributes, tc.attrs...),
-				expectedMethod:     MethodConnExec,
+				method:             MethodConnExec,
 				noParentSpan:       tc.noParentSpan,
 				ctx:                mc.execContextCtx,
+				attributesGetter:   tc.attributesGetter,
+				query:              query,
+				args:               args,
 			})
 
 			assert.Equal(t, 1, mc.execContextCount)
@@ -272,7 +292,9 @@ func TestOtConn_ExecContext(t *testing.T) {
 }
 
 func TestOtConn_QueryContext(t *testing.T) {
-	expectedAttrs := []attribute.KeyValue{semconv.DBStatementKey.String("query")}
+	query := "query"
+	args := []driver.NamedValue{{Value: "foo"}}
+	expectedAttrs := []attribute.KeyValue{semconv.DBStatementKey.String(query)}
 
 	for _, omitConnQuery := range []bool{true, false} {
 		var testname string
@@ -282,11 +304,12 @@ func TestOtConn_QueryContext(t *testing.T) {
 
 		t.Run(testname, func(t *testing.T) {
 			testCases := []struct {
-				name         string
-				error        bool
-				noParentSpan bool
-				disableQuery bool
-				attrs        []attribute.KeyValue
+				name             string
+				error            bool
+				noParentSpan     bool
+				disableQuery     bool
+				attrs            []attribute.KeyValue
+				attributesGetter AttributesGetter
 			}{
 				{
 					name:  "no error",
@@ -306,6 +329,11 @@ func TestOtConn_QueryContext(t *testing.T) {
 					noParentSpan: true,
 					attrs:        expectedAttrs,
 				},
+				{
+					name:             "with attribute getter",
+					attributesGetter: getDummyAttributesGetter(),
+					attrs:            expectedAttrs,
+				},
 			}
 
 			for _, tc := range testCases {
@@ -316,10 +344,11 @@ func TestOtConn_QueryContext(t *testing.T) {
 					// New conn
 					cfg.SpanOptions.DisableQuery = tc.disableQuery
 					cfg.SpanOptions.OmitConnQuery = omitConnQuery
+					cfg.AttributesGetter = tc.attributesGetter
 					mc := newMockConn(tc.error)
 					otelConn := newConn(mc, cfg)
 
-					rows, err := otelConn.QueryContext(ctx, "query", nil)
+					rows, err := otelConn.QueryContext(ctx, query, args)
 					if tc.error {
 						require.Error(t, err)
 					} else {
@@ -335,10 +364,13 @@ func TestOtConn_QueryContext(t *testing.T) {
 						parentSpan:         dummySpan,
 						error:              tc.error,
 						expectedAttributes: append(cfg.Attributes, tc.attrs...),
-						expectedMethod:     MethodConnQuery,
+						method:             MethodConnQuery,
 						noParentSpan:       tc.noParentSpan,
 						ctx:                mc.queryContextCtx,
 						omitSpan:           omitConnQuery,
+						attributesGetter:   tc.attributesGetter,
+						query:              query,
+						args:               args,
 					})
 
 					assert.Equal(t, 1, mc.queryContextCount)
@@ -368,7 +400,8 @@ func TestOtConn_QueryContext(t *testing.T) {
 }
 
 func TestOtConn_PrepareContext(t *testing.T) {
-	expectedAttrs := []attribute.KeyValue{semconv.DBStatementKey.String("query")}
+	query := "query"
+	expectedAttrs := []attribute.KeyValue{semconv.DBStatementKey.String(query)}
 
 	for _, omitConnPrepare := range []bool{true, false} {
 		var testname string
@@ -378,11 +411,12 @@ func TestOtConn_PrepareContext(t *testing.T) {
 
 		t.Run(testname, func(t *testing.T) {
 			testCases := []struct {
-				name         string
-				error        bool
-				noParentSpan bool
-				disableQuery bool
-				attrs        []attribute.KeyValue
+				name             string
+				error            bool
+				noParentSpan     bool
+				disableQuery     bool
+				attrs            []attribute.KeyValue
+				attributesGetter AttributesGetter
 			}{
 				{
 					name:  "no error",
@@ -402,6 +436,11 @@ func TestOtConn_PrepareContext(t *testing.T) {
 					noParentSpan: true,
 					attrs:        expectedAttrs,
 				},
+				{
+					name:             "with attribute getter",
+					attributesGetter: getDummyAttributesGetter(),
+					attrs:            expectedAttrs,
+				},
 			}
 
 			for _, tc := range testCases {
@@ -412,10 +451,11 @@ func TestOtConn_PrepareContext(t *testing.T) {
 					// New conn
 					cfg.SpanOptions.DisableQuery = tc.disableQuery
 					cfg.SpanOptions.OmitConnPrepare = omitConnPrepare
+					cfg.AttributesGetter = tc.attributesGetter
 					mc := newMockConn(tc.error)
 					otelConn := newConn(mc, cfg)
 
-					stmt, err := otelConn.PrepareContext(ctx, "query")
+					stmt, err := otelConn.PrepareContext(ctx, query)
 					if tc.error {
 						require.Error(t, err)
 					} else {
@@ -431,10 +471,12 @@ func TestOtConn_PrepareContext(t *testing.T) {
 						parentSpan:         dummySpan,
 						error:              tc.error,
 						expectedAttributes: append(cfg.Attributes, tc.attrs...),
-						expectedMethod:     MethodConnPrepare,
+						method:             MethodConnPrepare,
 						noParentSpan:       tc.noParentSpan,
 						ctx:                mc.prepareContextCtx,
 						omitSpan:           omitConnPrepare,
+						attributesGetter:   tc.attributesGetter,
+						query:              query,
 					})
 
 					assert.Equal(t, 1, mc.prepareContextCount)
@@ -453,9 +495,10 @@ func TestOtConn_PrepareContext(t *testing.T) {
 
 func TestOtConn_BeginTx(t *testing.T) {
 	testCases := []struct {
-		name         string
-		error        bool
-		noParentSpan bool
+		name             string
+		error            bool
+		noParentSpan     bool
+		attributesGetter AttributesGetter
 	}{
 		{
 			name: "no error",
@@ -468,6 +511,10 @@ func TestOtConn_BeginTx(t *testing.T) {
 			name:         "no parent span",
 			noParentSpan: true,
 		},
+		{
+			name:             "with attribute getter",
+			attributesGetter: getDummyAttributesGetter(),
+		},
 	}
 
 	for _, tc := range testCases {
@@ -476,6 +523,7 @@ func TestOtConn_BeginTx(t *testing.T) {
 			ctx, cfg, sr, dummySpan := prepareTraces(t, tc.noParentSpan)
 
 			// New conn
+			cfg.AttributesGetter = tc.attributesGetter
 			mc := newMockConn(tc.error)
 			otelConn := newConn(mc, cfg)
 
@@ -495,9 +543,10 @@ func TestOtConn_BeginTx(t *testing.T) {
 				parentSpan:         dummySpan,
 				error:              tc.error,
 				expectedAttributes: cfg.Attributes,
-				expectedMethod:     MethodConnBeginTx,
+				method:             MethodConnBeginTx,
 				noParentSpan:       tc.noParentSpan,
 				ctx:                mc.beginTxCtx,
+				attributesGetter:   tc.attributesGetter,
 			})
 
 			assert.Equal(t, 1, mc.beginTxCount)
@@ -523,9 +572,10 @@ func TestOtConn_ResetSession(t *testing.T) {
 
 		t.Run(testname, func(t *testing.T) {
 			testCases := []struct {
-				name         string
-				error        bool
-				noParentSpan bool
+				name             string
+				error            bool
+				noParentSpan     bool
+				attributesGetter AttributesGetter
 			}{
 				{
 					name: "no error",
@@ -538,6 +588,10 @@ func TestOtConn_ResetSession(t *testing.T) {
 					name:         "no parent span",
 					noParentSpan: true,
 				},
+				{
+					name:             "with attribute getter",
+					attributesGetter: getDummyAttributesGetter(),
+				},
 			}
 
 			for _, tc := range testCases {
@@ -547,6 +601,7 @@ func TestOtConn_ResetSession(t *testing.T) {
 
 					// New conn
 					cfg.SpanOptions.OmitConnResetSession = omitResetSession
+					cfg.AttributesGetter = tc.attributesGetter
 					mc := newMockConn(tc.error)
 					otelConn := newConn(mc, cfg)
 
@@ -566,10 +621,11 @@ func TestOtConn_ResetSession(t *testing.T) {
 						parentSpan:         dummySpan,
 						error:              tc.error,
 						expectedAttributes: cfg.Attributes,
-						expectedMethod:     MethodConnResetSession,
+						method:             MethodConnResetSession,
 						noParentSpan:       tc.noParentSpan,
 						ctx:                mc.resetSessionCtx,
 						omitSpan:           omitResetSession,
+						attributesGetter:   tc.attributesGetter,
 					})
 
 					assert.Equal(t, 1, mc.resetSessionCount)