+++ instruments.go
@@ -20,9 +20,6 @@ import (
 
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/instrument"
-	"go.opentelemetry.io/otel/metric/instrument/asyncfloat64"
-	"go.opentelemetry.io/otel/metric/instrument/asyncint64"
-	"go.opentelemetry.io/otel/metric/instrument/syncfloat64"
 	"go.opentelemetry.io/otel/metric/unit"
 )
 
@@ -31,25 +28,25 @@ const (
 )
 
 type dbStatsInstruments struct {
-	connectionMaxOpen                asyncint64.Gauge
-	connectionOpen                   asyncint64.Gauge
-	connectionWaitTotal              asyncint64.Counter
-	connectionWaitDurationTotal      asyncfloat64.Counter
-	connectionClosedMaxIdleTotal     asyncint64.Counter
-	connectionClosedMaxIdleTimeTotal asyncint64.Counter
-	connectionClosedMaxLifetimeTotal asyncint64.Counter
+	connectionMaxOpen                instrument.Int64ObservableGauge
+	connectionOpen                   instrument.Int64ObservableGauge
+	connectionWaitTotal              instrument.Int64ObservableCounter
+	connectionWaitDurationTotal      instrument.Float64ObservableCounter
+	connectionClosedMaxIdleTotal     instrument.Int64ObservableCounter
+	connectionClosedMaxIdleTimeTotal instrument.Int64ObservableCounter
+	connectionClosedMaxLifetimeTotal instrument.Int64ObservableCounter
 }
 
 type instruments struct {
 	// The latency of calls in milliseconds
-	latency syncfloat64.Histogram
+	latency instrument.Float64Histogram
 }
 
 func newInstruments(meter metric.Meter) (*instruments, error) {
 	var instruments instruments
 	var err error
 
-	if instruments.latency, err = meter.SyncFloat64().Histogram(
+	if instruments.latency, err = meter.Float64Histogram(
 		strings.Join([]string{namespace, "latency"}, "."),
 		instrument.WithDescription("The latency of calls in milliseconds"),
 		instrument.WithUnit(unit.Milliseconds),
@@ -64,51 +61,51 @@ func newDBStatsInstruments(meter metric.Meter) (*dbStatsInstruments, error) {
 	var err error
 	subsystem := "connection"
 
-	if instruments.connectionMaxOpen, err = meter.AsyncInt64().Gauge(
+	if instruments.connectionMaxOpen, err = meter.Int64ObservableGauge(
 		strings.Join([]string{namespace, subsystem, "max_open"}, "."),
 		instrument.WithDescription("Maximum number of open connections to the database"),
 	); err != nil {
 		return nil, fmt.Errorf("failed to create connectionMaxOpen instrument, %v", err)
 	}
 
-	if instruments.connectionOpen, err = meter.AsyncInt64().Gauge(
+	if instruments.connectionOpen, err = meter.Int64ObservableGauge(
 		strings.Join([]string{namespace, subsystem, "open"}, "."),
 		instrument.WithDescription("The number of established connections both in use and idle"),
 	); err != nil {
 		return nil, fmt.Errorf("failed to create connectionOpen instrument, %v", err)
 	}
 
-	if instruments.connectionWaitTotal, err = meter.AsyncInt64().Counter(
-		strings.Join([]string{namespace, subsystem, "wait_total"}, "."),
+	if instruments.connectionWaitTotal, err = meter.Int64ObservableCounter(
+		strings.Join([]string{namespace, subsystem, "wait"}, "."),
 		instrument.WithDescription("The total number of connections waited for"),
 	); err != nil {
 		return nil, fmt.Errorf("failed to create connectionWaitTotal instrument, %v", err)
 	}
 
-	if instruments.connectionWaitDurationTotal, err = meter.AsyncFloat64().Counter(
-		strings.Join([]string{namespace, subsystem, "wait_duration_total"}, "."),
+	if instruments.connectionWaitDurationTotal, err = meter.Float64ObservableCounter(
+		strings.Join([]string{namespace, subsystem, "wait_duration"}, "."),
 		instrument.WithDescription("The total time blocked waiting for a new connection"),
 		instrument.WithUnit(unit.Milliseconds),
 	); err != nil {
 		return nil, fmt.Errorf("failed to create connectionWaitDurationTotal instrument, %v", err)
 	}
 
-	if instruments.connectionClosedMaxIdleTotal, err = meter.AsyncInt64().Counter(
-		strings.Join([]string{namespace, subsystem, "closed_max_idle_total"}, "."),
+	if instruments.connectionClosedMaxIdleTotal, err = meter.Int64ObservableCounter(
+		strings.Join([]string{namespace, subsystem, "closed_max_idle"}, "."),
 		instrument.WithDescription("The total number of connections closed due to SetMaxIdleConns"),
 	); err != nil {
 		return nil, fmt.Errorf("failed to create connectionClosedMaxIdleTotal instrument, %v", err)
 	}
 
-	if instruments.connectionClosedMaxIdleTimeTotal, err = meter.AsyncInt64().Counter(
-		strings.Join([]string{namespace, subsystem, "closed_max_idle_time_total"}, "."),
+	if instruments.connectionClosedMaxIdleTimeTotal, err = meter.Int64ObservableCounter(
+		strings.Join([]string{namespace, subsystem, "closed_max_idle_time"}, "."),
 		instrument.WithDescription("The total number of connections closed due to SetConnMaxIdleTime"),
 	); err != nil {
 		return nil, fmt.Errorf("failed to create connectionClosedMaxIdleTimeTotal instrument, %v", err)
 	}
 
-	if instruments.connectionClosedMaxLifetimeTotal, err = meter.AsyncInt64().Counter(
-		strings.Join([]string{namespace, subsystem, "closed_max_lifetime_total"}, "."),
+	if instruments.connectionClosedMaxLifetimeTotal, err = meter.Int64ObservableCounter(
+		strings.Join([]string{namespace, subsystem, "closed_max_lifetime"}, "."),
 		instrument.WithDescription("The total number of connections closed due to SetConnMaxLifetime"),
 	); err != nil {
 		return nil, fmt.Errorf("failed to create connectionClosedMaxLifetimeTotal instrument, %v", err)