+++ rows_test.go
@@ -171,8 +171,9 @@ func TestNewRows(t *testing.T) {
 
 		t.Run(testname, func(t *testing.T) {
 			testCases := []struct {
-				name         string
-				noParentSpan bool
+				name             string
+				noParentSpan     bool
+				attributesGetter AttributesGetter
 			}{
 				{
 					name: "default config",
@@ -180,6 +181,10 @@ func TestNewRows(t *testing.T) {
 				{
 					name:         "no parent span",
 					noParentSpan: true,
+				},
+				{
+					name:             "with attribute getter",
+					attributesGetter: getDummyAttributesGetter(),
 				},
 			}
 
@@ -190,6 +195,7 @@ func TestNewRows(t *testing.T) {
 
 					mr := newMockRows(false)
 					cfg.SpanOptions.OmitRows = omitRows
+					cfg.AttributesGetter = tc.attributesGetter
 
 					// New rows
 					rows := newRows(ctx, mr, cfg)
@@ -209,10 +215,11 @@ func TestNewRows(t *testing.T) {
 						parentSpan:         dummySpan,
 						error:              false,
 						expectedAttributes: cfg.Attributes,
-						expectedMethod:     MethodRows,
+						method:             MethodRows,
 						noParentSpan:       tc.noParentSpan,
 						spanNotEnded:       true,
 						omitSpan:           omitRows,
+						attributesGetter:   tc.attributesGetter,
 					})
 
 					assert.Equal(t, mr, rows.Rows)